# slooze-api-spec.json
{"openapi":"3.1.0","info":{"title":"Slooze Food Ordering API","version":"1.0.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_v1_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Read Users Me","operationId":"read_users_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/restaurants/":{"get":{"tags":["Restaurants"],"summary":"List Restaurants","operationId":"list_restaurants_api_v1_restaurants__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Restaurant"},"type":"array","title":"Response List Restaurants Api V1 Restaurants  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/restaurants/{restaurant_id}":{"get":{"tags":["Restaurants"],"summary":"Get Restaurant","operationId":"get_restaurant_api_v1_restaurants__restaurant_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"restaurant_id","in":"path","required":true,"schema":{"type":"integer","title":"Restaurant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestaurantWithMenu"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/restaurants/{restaurant_id}/menu":{"get":{"tags":["Restaurants"],"summary":"Get Restaurant Menu","operationId":"get_restaurant_menu_api_v1_restaurants__restaurant_id__menu_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"restaurant_id","in":"path","required":true,"schema":{"type":"integer","title":"Restaurant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuItem"},"title":"Response Get Restaurant Menu Api V1 Restaurants  Restaurant Id  Menu Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/orders/":{"get":{"tags":["Orders"],"summary":"List Orders","operationId":"list_orders_api_v1_orders__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrderResponse"},"type":"array","title":"Response List Orders Api V1 Orders  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Orders"],"summary":"Create New Order","operationId":"create_new_order_api_v1_orders__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/orders/{order_id}":{"get":{"tags":["Orders"],"summary":"Get Order","operationId":"get_order_api_v1_orders__order_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Orders"],"summary":"Cancel","operationId":"cancel_api_v1_orders__order_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/orders/{order_id}/items":{"post":{"tags":["Orders"],"summary":"Add Order Items","operationId":"add_order_items_api_v1_orders__order_id__items_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemCreate"},"title":"Items"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/orders/{order_id}/checkout":{"post":{"tags":["Orders"],"summary":"Checkout","operationId":"checkout_api_v1_orders__order_id__checkout_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCheckout"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/payment-methods/":{"get":{"tags":["Payment Methods"],"summary":"List Payment Methods","operationId":"list_payment_methods_api_v1_payment_methods__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PaymentMethodResponse"},"type":"array","title":"Response List Payment Methods Api V1 Payment Methods  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Payment Methods"],"summary":"Create Payment","operationId":"create_payment_api_v1_payment_methods__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentMethodCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentMethodResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/payment-methods/{payment_id}":{"put":{"tags":["Payment Methods"],"summary":"Update Payment","operationId":"update_payment_api_v1_payment_methods__payment_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"payment_id","in":"path","required":true,"schema":{"type":"integer","title":"Payment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentMethodUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentMethodResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Payment Methods"],"summary":"Delete Payment","operationId":"delete_payment_api_v1_payment_methods__payment_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"payment_id","in":"path","required":true,"schema":{"type":"integer","title":"Payment Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_api_v1_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_v1_auth_login_post"},"Country":{"type":"string","enum":["INDIA","AMERICA"],"title":"Country"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MenuItem":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"price":{"type":"number","title":"Price"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"id":{"type":"integer","title":"Id"},"restaurant_id":{"type":"integer","title":"Restaurant Id"},"is_available":{"type":"boolean","title":"Is Available"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["name","price","id","restaurant_id","is_available","created_at"],"title":"MenuItem"},"OrderCheckout":{"properties":{"payment_method_id":{"type":"integer","title":"Payment Method Id"}},"type":"object","required":["payment_method_id"],"title":"OrderCheckout"},"OrderCreate":{"properties":{"restaurant_id":{"type":"integer","title":"Restaurant Id"},"items":{"items":{"$ref":"#/components/schemas/OrderItemCreate"},"type":"array","title":"Items"}},"type":"object","required":["restaurant_id","items"],"title":"OrderCreate"},"OrderItemCreate":{"properties":{"menu_item_id":{"type":"integer","title":"Menu Item Id"},"quantity":{"type":"integer","title":"Quantity"}},"type":"object","required":["menu_item_id","quantity"],"title":"OrderItemCreate"},"OrderItemResponse":{"properties":{"id":{"type":"integer","title":"Id"},"menu_item_id":{"type":"integer","title":"Menu Item Id"},"quantity":{"type":"integer","title":"Quantity"},"price_at_order":{"type":"number","title":"Price At Order"},"subtotal":{"type":"number","title":"Subtotal"}},"type":"object","required":["id","menu_item_id","quantity","price_at_order","subtotal"],"title":"OrderItemResponse"},"OrderResponse":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"restaurant_id":{"type":"integer","title":"Restaurant Id"},"status":{"$ref":"#/components/schemas/OrderStatus"},"total_amount":{"type":"number","title":"Total Amount"},"country":{"$ref":"#/components/schemas/Country"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"order_items":{"items":{"$ref":"#/components/schemas/OrderItemResponse"},"type":"array","title":"Order Items","default":[]}},"type":"object","required":["id","user_id","restaurant_id","status","total_amount","country","created_at","updated_at"],"title":"OrderResponse"},"OrderStatus":{"type":"string","enum":["PENDING","CONFIRMED","CANCELLED"],"title":"OrderStatus"},"PaymentMethodCreate":{"properties":{"method_type":{"$ref":"#/components/schemas/PaymentMethodType"},"details":{"type":"string","title":"Details"},"is_default":{"type":"boolean","title":"Is Default","default":false}},"type":"object","required":["method_type","details"],"title":"PaymentMethodCreate"},"PaymentMethodResponse":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"method_type":{"$ref":"#/components/schemas/PaymentMethodType"},"details":{"type":"string","title":"Details"},"is_default":{"type":"boolean","title":"Is Default"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","user_id","method_type","details","is_default","created_at"],"title":"PaymentMethodResponse"},"PaymentMethodType":{"type":"string","enum":["CREDIT_CARD","DEBIT_CARD","UPI","WALLET"],"title":"PaymentMethodType"},"PaymentMethodUpdate":{"properties":{"details":{"type":"string","title":"Details"},"is_default":{"type":"boolean","title":"Is Default","default":false}},"type":"object","required":["details"],"title":"PaymentMethodUpdate"},"Restaurant":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"cuisine_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cuisine Type"},"country":{"$ref":"#/components/schemas/Country"},"id":{"type":"integer","title":"Id"},"rating":{"type":"number","title":"Rating"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["name","country","id","rating","is_active","created_at"],"title":"Restaurant"},"RestaurantWithMenu":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"cuisine_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cuisine Type"},"country":{"$ref":"#/components/schemas/Country"},"id":{"type":"integer","title":"Id"},"rating":{"type":"number","title":"Rating"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"menu_items":{"items":{"$ref":"#/components/schemas/MenuItem"},"type":"array","title":"Menu Items","default":[]}},"type":"object","required":["name","country","id","rating","is_active","created_at"],"title":"RestaurantWithMenu"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"User":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"role":{"$ref":"#/components/schemas/UserRole"},"country":{"$ref":"#/components/schemas/Country"},"id":{"type":"integer","title":"Id"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["email","full_name","role","country","id","is_active","created_at"],"title":"User"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"role":{"$ref":"#/components/schemas/UserRole"},"country":{"$ref":"#/components/schemas/Country"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","full_name","role","country","password"],"title":"UserCreate"},"UserRole":{"type":"string","enum":["ADMIN","MANAGER","MEMBER"],"title":"UserRole"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}